Bugfix: interval.min(d) is long

+++ src/main/java/net/imglib2/algorithm/blocks/convolve/ConvolveProcessors.java
@@ -193,7 +193,7 @@ class ConvolveProcessors
 			assert interval.numDimensions() == n;
 
 			Arrays.setAll( destSize, d -> ( int ) interval.dimension( d ) );
-			Arrays.setAll( sourcePos, d -> ( int ) interval.min( d ) + kernelOffsets[ d ] );
+			Arrays.setAll( sourcePos, d -> interval.min( d ) + kernelOffsets[ d ] );
 
 			aux0Length = 0;
 			aux1Length = 0;
